requirements.txt
aiogram==3.2.0
python-dotenv==1.0.0
pytz==2023.3
mysql-connector-python==8.2.0
numpy==1.24.3
pandas==2.0.1
openpyxl==3.1.2
